HeadphoneKey: Press keys from `_key_buffer` map
Take the headphone keys in `_key_buffer` and get the associated map
action by calling into the Rust library. If the map action has
`MapKind::Map`, then we need to simulate key presses of the characters
in the `action` field.

Works, but it's a little finicky. Maybe I'm including the null byte in
the loop?