HeadphoneKey(runAction): Take `strlen` out of loop
This appears to fix the weirdness I was having in
815bf3e8601b6d0cfb0296c6f2c302aaf5059b23, where extra characters were
getting sent to `simpleKeyPressWithKey:` even though there was only 1
element (the correct number of elements) in the `action` array/string.

I guess before the length was getting recalculated on each iteration of
the loop.

Previously, this would result in the Vim `x` command getting entered
just after the map action (for example `j`).

Now it appears to work correctly.