HeadphoneKey(runAction): Fix EXC_BAD_ACCESS on undefined map
If no map was defined for a headphone button, the fields in the
`CKeyActionResult` `result` would be null pointers, causing the bad
access exception.

Before doing anything, check that the pointer points to something real.

Also, fix the `kind` check to use the `ActionKind_Map` enum now that the
`dome-key-map` library has been updated with the new definition of
`CKeyActionResult`.