LicenseHandler: Validate license files
Fill in the `check` method to have it validate a license file. If the
file doesn't validate, prints an error message.

A bunch of supporting methods to check if the license file at the
program's prescribed path exists.

Use AquaticPrime's CoreFoundation functions to validate the license
file. Include my public key, generated by the "AquaticPrime Developer"
app, with the CFStringAppend calls copied from the program's generated
"Obfuscated Public Key" text box.

Run the validation check on launch in `main.m`.

Add an `eprintf` macro to print to stderr with a prepended program
identifier.