LicenseHandler(addLicense:): Delete license file instead of trashing it
The trash/recycle method wasn't working because it was asynchronous. As
a result, the file never got moved.

Using `removeItemAtURL:error:` does work, and is available starting in
OS 10.6. But I still don't like the idea of forcibly deleting files.
What if the user licenses the program and deletes their license file,
then somehow gets their only local copy deleted?

Need to work out a different way of doing this.