LicenseHandler: Use current date string in file moved to trash
To prevent errors resulting from trying to move the license file to the
trash when a file with the same name already exists there, use a custom
file name with a time stamp.

Instead of naming the trashed file "license.plist", prepend "dome-key-"
to the filename to make it clear to users that the file came from
DomeKey. Append the current date in the format "2018-10-26-19.28.13"
before the file extension.