LicenseHandler(addLicense:): Delete license file at the start
Previously, you couldn't replace the existing license file using the
`--license` flag. If you somehow got an invalid license installed, you'd
get an error message asking you to add your license again:

    "Try adding your license again using the `--license` flag."

But this wouldn't work to replace the existing license. The copy
operation would fail complaining about the file already existing. The
behaviour of the method now is to always replace the license file, just
in case it's necessary.

We use the existing "move to trash" method, and make sure we ignore
"does not exist" errors, because typically when you run `--license`,
you're not going to have a `license.plist` file installed at the start,
so for our purposes, that's not an error.