HeadphoneKey: Play audio only with the `--audio` command line flag
Allow users to decide whether interface audio (mode activated &
deactivated) should play.

The flag value is in a `Config`, so to make it easier to pass both the
audio flag and the timeout config value to `HeadphoneKey`, rename
`initWithTimeout:` to `initWithConfig:`.

Only allocate a `Sounds` instance if `--audio` is passed, because
there's no reason to allocate memory for it otherwise.

Add a `_play_audio` flag as a static variable to make it accessible to
the `on_mode_change()` function, just like we did for `sounds_inst`.

In `on_mode_change()`, don't play any audio if `_play_audio` is on.