Add a skeleton Homebrew formula
The start of a Homebrew formula for the program. Needs a bunch of fields
filled out. For now, we're mostly getting it to work as an input
template for an accompanying generator script. The script will fill in
some values automatically, including the plist file, so that the launchd
plist can be activated using Homebrew's services.