Makefile: Use project-local build directory
Set 'DerivedData' path in `xcodebuild` to give us a local build
directory.

Doing this because when the 'Debug', or 'Release' directories in
'Build/Products/' don't exist, Make doesn't build our recipes. Make
seems to prefer local files, and I don't imagine the wildcard in the
path does us any favours.

When I changed the DerivedData path, I ended up with this error:

    In file included from DomeKey/DomeKey/HeadphoneKey.m:9:
    DomeKey/DomeKey/HeadphoneKey.h:10:9: fatal error:
          'DDHidLib/DDHidAppleMikey.h' file not found
    #import <DDHidLib/DDHidAppleMikey.h>
            ^~~~~~~~~~~~~~~~~~~~~~~~~~~~

Couldn't figure out what the problem was, and what changed there when I
changed the DerivedData directory. So I just fixed it in the only way I
could think of: adding the DDHidLib 'lib' directory to
`HEADER_SEARCH_PATHS` and updating the `#import` to point to the new
location.

We also create a release target in this commit, which gives us a release
build of the program.