HeadphoneKeyEventBluetooth: Initial thought with MPRemoteCommandCenter
Turns out that `MPRemoteCommandCenter` seems to work for the 'play'
button, but it has no interface for volume buttons. I remember
discovering the volume button problem when I last looked into it several
months ago, but since forgot.

This was an attempt at replicating the method in:
https://github.com/russellhancox/btrcd/blob/c3e375f/btrcd/AppDelegate.m
by setting up an `MPRemoteCommandCenter` and calling
`addTargetWithHandler:` on each button to invoke a custom handler.

Committing this for reference. In order to get volume button
functionality, we'll have to dive into lower level Bluetooth events with
AVRCP. Seems that newer headphones might even use something newer than
AVRCP, so I wonder if we'll have to use multiple Bluetooth remote
control profiles.

A bit of a pain, as there doesn't seem to be a pre-built interface for
us to do this, but I'm learning more about interacting with IOBluetooth.
Also, `MPRemoteCommandCenter` is an OS X 10.12 Sierra API, so if there's
something we can use that's available on earlier OSes, that's a plus.