HeadphoneKeyEventBluetooth: Listen for Bluetooth device pairing
Register a delegate that gets called when Bluetooth devices are paired
with the Mac.

We now need to determine if the paired device is a pair of headphones
(ideally if we can figure out media remote control capability that would
be helpful), and then connect to them over L2CAP to get AVRCP messages.

The `devicePairingFinished:error:` delegate method was a test. Leaving
it in as a reminder that we also need to implement functionality that
looks for already-connected Bluetooth headphones when DomeKey launches.