HeadphoneKeyEventBluetooth: Add L2CAP connection
Open a channel with Bluetooth headphones over L2CAP. The goal is to get
button press events from the channel.

Unfortunately, I seem to only be able to get press/release events from
the "play" button but not the volume buttons. Weird. Not sure what's
going on there yet. Is there a different channel for volume controls?

Came across `kBluetoothSDPAttributeIdentifierRemoteAudioVolumeControl`
in `BluetoothAssignedNumbers.h`. Is that interesting to us?